+++ packages/pdf-viewer/src/PdfViewer.tsx
@@ -1,4 +1,4 @@
-import { useEffect, useState } from 'react';
+import { forwardRef, useEffect, useState } from 'react';
 import cx from 'classnames';
 
 import './index.css';
@@ -11,13 +11,10 @@ export interface IPdfViewerProps {
   style?: React.CSSProperties;
 }
 
-const PdfViewer = ({
-  blob,
-  width = 600,
-  height = 800,
-  className = '',
-  style = {},
-}: IPdfViewerProps) => {
+const PdfViewer = forwardRef<HTMLIFrameElement, IPdfViewerProps>(function PdfViewer(
+  { blob, width = 600, height = 800, className = '', style = {} }: IPdfViewerProps,
+  ref,
+) {
   const [iframeSrc, setIframeSrc] = useState('');
 
   useEffect(() => {
@@ -31,12 +28,12 @@ const PdfViewer = ({
   return (
     <div style={{ ...style, width: width, height: height }} className={cx('pdf-viewer', className)}>
       {blob && (
-        <iframe src={iframeSrc} style={{ width: '100%', height: '100%' }}>
+        <iframe ref={ref} src={iframeSrc} style={{ width: '100%', height: '100%' }}>
           <p>您的浏览器不支持 iframe 标签，请从列表中下载预览</p>
         </iframe>
       )}
     </div>
   );
-};
+});
 
 export default PdfViewer;